refactor: Move validate_filename func to helper
